Serialize and deserialize chunks directly into the stream. Start making a networked VoxelEditor. Reworking the networking system

+++ Clunker/Voxels/Voxel.cs
@@ -40,7 +40,7 @@ namespace Clunker.Voxels
 
         public override int GetHashCode()
         {
-            return (int)Data;
+            return Data;
         }
     }
